+++ search-films/src/components/shared/CssUtils.js
@@ -1,0 +1,6 @@
+class CssUtils {
+	static addPxToCssProperty(property) {
+		return property + 'px';
+	}
+}
+export default CssUtils;