+++ search-films/CHANGELOG.md
@@ -1,8 +1,13 @@
 #### 0.0.1 (2024-08-05)
 
+##### Chores
+
+- fix CHANGELOG.md (c77c22d8)
+
 ##### New Features
 
-- add Title component (02742be1)
+- Add Paragraph component (8d5ea4fb)
+- add Title component (38198954)
 - add Button component (037a7129)
 
 #### 0.0.0 (2024-08-05)